requirements.txt
# OSM and geospatial
osmnx>=1.9.0
overpy>=0.5
networkx>=3.2
geopandas>=0.14.0
shapely>=2.0.0
geopy>=2.4.0

# Scraping
playwright>=1.40.0
scrapy-rotating-proxies>=0.6.0

# Scheduling and storage
apscheduler>=3.10.0

# Configuration and utilities
pydantic>=2.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
